Moved HtmlEntity to Plivo.Utilities namespace
